requirements.txt
azure-functions
ffmpeg-python
pydub